Sashank/fixes

+++ server/controllers/listingController.js
@@ -1,7 +1,6 @@
 const Listing = require('../models/listingModel');
 const User = require('../models/userModel');
 const Pool = require('../models/poolModel');
-const data = require('./initalData.js');
 const mongoose = require('mongoose');
 
 const getListings = async (req, res) => {
@@ -10,8 +9,9 @@ const getListings = async (req, res) => {
 };
 
 const getFilteredListings = async (req, res) => {
-  Listing.createIndexes({ price: 1 });
-  Listing.createIndexes({ createdAt: 1 });
+  await Listing.createIndexes({ price: 1 });
+  await Listing.createIndexes({ createdAt: 1 });
+  await Listing.createIndexes({ '$**': 'text' });
 
   let queryDecoded = '';
   let filterObj = {};
@@ -91,7 +91,6 @@ const getFilteredListings = async (req, res) => {
   filterObj = filterArr.reduce((acc, curr, i) => ({
     ...acc,
     [`${Object.keys(curr)[0]}`]: Object.values(curr)[0],
-    //Ref: https://www.appsloveworld.com/nodejs/100/321/how-do-i-add-multiple-optional-parameters-in-express-in-same-route
   }));
   console.log(filterObj);
 
@@ -112,7 +111,7 @@ const addListing = async (req, res) => {
     throw new Error('Please specify a name, address, price, and email');
   }
 
-  const images = req.files.map((file) => file.location); // Retrieve the file paths of all the uploaded images
+  const images = req.files.map((file) => file.location);
 
   const listing = await Listing.create({
     ...req.body,
@@ -132,7 +131,7 @@ const updateListing = async (req, res) => {
     throw new Error('listing not found');
   }
 
-  const images = req.files.map((file) => file.location); // Retrieve the file paths of all the uploaded images
+  const images = req.files.map((file) => file.location);
 
   const updatedListing = await Listing.findByIdAndUpdate(
     req.params.id,
@@ -149,6 +148,10 @@ const deleteListing = async (req, res) => {
   if (!listing) {
     res.status(400);
     throw new Error('Listing not found');
+  }
+  if (listing.status === 'Sold') {
+    res.status(400);
+    throw new Error('Cannot delete a sold listing');
   }
   await Listing.deleteOne({ _id: req.params.id });
   res.status(200).json({ id: req.params.id });
@@ -189,8 +192,6 @@ const sellListing = async (req, res) => {
   await Pool.deleteOne({ _id: req.params.poolId });
   res.status(200).json(listing);
 };
-
-// TODO : Filtering listings
 
 module.exports = {
   getListings,
